pythonA2A/pyproject.toml
[project]
name = "a2a-clients"
version = "0.1.0"
description = "A2A Client Sample"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
  "google-adk[a2a]",
  "a2a-sdk",
  "python-dotenv",
  "agent-framework-a2a>=1.0.0b260311",
  "beeai-framework[a2a]>=0.1.78",
  "agentstack-sdk>=0.6.2",
  "watchfiles>=1.1.1",
]
